requirements.txt
diffusers>=0.37.1
transformers
safetensors
numpy
pillow
